+++ R/CallbackBatchTuning.R
@@ -178,12 +178,6 @@ CallbackBatchTuning = R6Class(
 #'  Called in `OptimInstance$assign_result()`.
 #'  The functions must have two arguments named `callback` and `context`.
 #'  The final result `instance$result` is available in `context`.
-#' @param on_result (`function()`)\cr
-#'  Deprecated.
-#'  Use `on_result_end` instead.
-#'  Stage called after the result is written.
-#'  Called in `OptimInstance$assign_result()`.
-#'  The functions must have two arguments named `callback` and `context`.
 #' @param on_optimization_end (`function()`)\cr
 #'  Stage called at the end of the optimization.
 #'  Called in `Optimizer$optimize()`.
@@ -208,7 +202,6 @@ callback_batch_tuning = function(
   on_tuning_result_begin = NULL,
   on_result_begin = NULL,
   on_result_end = NULL,
-  on_result = NULL,
   on_optimization_end = NULL
 ) {
   stages = discard(
@@ -227,7 +220,6 @@ callback_batch_tuning = function(
         on_tuning_result_begin,
         on_result_begin,
         on_result_end,
-        on_result,
         on_optimization_end
       ),
       c(
@@ -244,21 +236,11 @@ callback_batch_tuning = function(
         "on_tuning_result_begin",
         "on_result_begin",
         "on_result_end",
-        "on_result",
         "on_optimization_end"
       )
     ),
     is.null
   )
-
-  if ("on_result" %in% names(stages)) {
-    if ("on_result_end" %in% names(stages)) {
-      stopf("Both `on_result` and `on_result_end` are supplied. `on_result` is deprecated, use only `on_result_end`.")
-    }
-    .Deprecated(old = "on_result", new = "on_result_end")
-    stages$on_result_end = stages$on_result
-    stages$on_result = NULL
-  }
 
   walk(stages, function(stage) assert_function(stage, args = c("callback", "context")))
   callback = CallbackBatchTuning$new(id, label, man)